Added new fields to contact (emaill, address1, address2, city, state, zip), updated methods to account for these changes,
Updated updatecontact method to check to see if you even want to update any of the values, and if you dont you dont have to enter the value in again.

Made it so that it will only display values that are there, if you dont have an address listed it wont show anything, instead of list listing a bunch of blanks.